wifi_app/requirements.txt
PySide6
speedtest-cli
matplotlib
scapy
pandas
requests
psutil
